refactor(rewrite): migrate tickets command to Listener class, subclass View and add callbacks
