+++ func/classes/Func.py
@@ -4,8 +4,7 @@ from __main__ import bot
 import asyncio
 from concurrent.futures import ThreadPoolExecutor
 from datetime import datetime
-import discord 
-from discord import ButtonStyle, SelectOption
+import discord
 import requests
 
 from func.utils.discord_utils import create_ticket, create_transcript, name_grabber, log_event
@@ -365,57 +364,3 @@ class Func:
                     embeds.append(embed)
 
             return embeds
-            
-    async def tickets():
-        embed = discord.Embed(title="Tickets",
-                            description="""Tickets can be created for any of the following reasons:
-                                        > Do-not-kick-list Application
-                                        > Discord Nick/Role Change
-                                        > Problems/Queries/Complaints
-                                        > Player Report
-                                        > Milestone
-                                        > Staff Application
-                                        > Event
-                                        > Other
-                                        Once you have created a ticket by clicking the button, you will be linked to your ticket\n
-                                        The bot will ask you to choose the reason behind the creation of your ticket from a given list. Choose the appropriate reason and then proceed!\n
-                                        Once you have created your ticket, staff will respond within 24 hours.""",
-                            color=neutral_color)
-
-        embed.add_field(name="Do-not-kick-list Application",
-                        value="You  must have a valid reason for applying and also meet the DNKL requiremnets.\n"
-                              "Accepted Reasons:\n"
-                              "> School\n"
-                              "> Medical Reasons\n"
-                              "> Situations out of your control\n"
-                              "> Vacation\n\n"
-                              "If your account is banned, it may be temporarily kicked until unbanned.",
-                        inline=False)
-
-        embed.add_field(name="Player Report",
-                        value="When reporting a player, you're expected to explain the situation in maximum detail. Providing the following is considered the bare minimum:\n"
-                              "> Username of the accused\n"
-                              "> Explanantion of the offense\n"
-                              "> Time of offense\n"
-                              "> Proof of offense\n"
-                              "If you wish to report a staff member, please DM the acting guild master with your report.",
-                        inline=False)
-
-        embed.add_field(name="Milestone",
-                        value="You'll be prompted to present the milestone you've achieved and proof of its occurence. "
-                              "Staff will review your milestone and if accepted, will be include it in the next week's milestone post!",
-                        inline=False)
-
-        embed.add_field(name="Staff Application",
-                        value="After you're done with your application, the staff team will review your it and make a decision to accept or deny it.",
-                        inline=False)
-
-        embed.set_thumbnail(
-            url=f"https://images-ext-1.discordapp.net/external/ziYSZZe7dPyKDYLxA1s2jqpKi-kdCvPFpPaz3zft-wo/%3Fwidth%3D671%26height%3D671/https/media.discordapp.net/attachments/523227151240134664/803843877999607818/misc.png")
-
-        image = await get_gtop("https://media.discordapp.net/attachments/650248396480970782/873866686049189898/tickets.jpg")
-
-        view = View()
-        view.add_item(Button(style=ButtonStyle.blurple, custom_id="ticketer", label="✉️Create Ticket"))
-
-        return image, embed, view
