+++ func/cogs/menus.py
@@ -1,7 +1,6 @@
 from discord.ext import commands
 
 from func.classes.Listener import Listener
-from func.classes.Func import Func
 
 
 class Menus(commands.Cog, command_attrs=dict(hidden=True), name="menus"):
@@ -19,7 +18,7 @@ class Menus(commands.Cog, command_attrs=dict(hidden=True), name="menus"):
     @commands.is_owner()
     async def tickets(self, ctx):
         """Send a ticket help embed!"""
-        image, embed, view = await Func.tickets()
+        image, embed, view = await Listener.tickets()
         await ctx.send(file=image, embed=embed, view=view)
